requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt>=5.5.1
python-decouple==3.8
Pillow>=10.2.0
setuptools>=65.0.0
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0

